Remove 'notification' mention from System Permission
It seems that setting the 'infoPlist' for notification is not needed. (https://github.com/expo/expo/issues/1040)
This mention that I'm deleting seems to indicate that it is needed...
It took me a while to figure it out and it seems that mentioning this here will only create more confusion as it has created for me.